Refactor: Changed attribute names acc. to PEP8.
Many attributes were poorly named. This was fixed by renaming and correcting all attributes according to PEP8. A preliminary conversion function for LH to RH conversion of Rotations and Vectors was created. The inverse is still left and is being looked upon. It is recommended to go through the API function definitions once again.